backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-rest==0.8.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
fonttools==4.59.1
groq==0.31.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.11
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
